moving code to src/
