+++ tsconfig.eslint.json
@@ -1,6 +1,7 @@
 {
   "extends": "./tsconfig.json",
   "include": [
+    "src/*.ts",
     "index.ts",
     "__mocks__/**/*.ts",
     "__tests__/**/*.ts",
